backend/requirements.txt
# AI MediScan Backend - Production Dependencies
# Python 3.11+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Cloud & AI
google-cloud-aiplatform==1.42.0
google-generativeai==0.3.2
google-cloud-storage==2.14.0
firebase-admin==6.4.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security & Auth
PyJWT==2.8.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Rate Limiting & Caching
slowapi==0.1.9
redis==5.0.1

# HTTP & WebSocket
httpx==0.26.0
websockets==12.0

# Monitoring & Logging
structlog==24.1.0
sentry-sdk==1.40.0

# Utilities
python-dateutil==2.8.2
pillow==10.2.0
tenacity==8.2.3

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0


